refactor(experiments): the reusable report pieces move out of old/
The Results screen needs the Chart.js options, the interactive legend plugin, the
p-chart wrapper, the headless details table and the winner/promoted/chart-data
helpers. All of them live under old/, which is frozen and slated for deletion, so
they are relocated to shared/ and old/ is repointed at the new paths rather than
duplicating them: one source of truth, and old/ stays deletable.

Only import lines change under old/, so its behaviour is untouched and its specs
still pass unmodified. Three importers the plan had not accounted for turned up
in the Configure legacy screen and needed the same repoint.

getConversionRate, getConversionRateRage, getBayesianVariantResult and
getProbabilityToBeBest stay where they are: they are not needed outside old/, and
the dependency direction has to remain old/ -> shared/, never the reverse.
