+++ core-web/libs/portlets/dot-experiments/portlet/src/lib/old/dot-experiments-reports/components/dot-experiments-report-daily-details/dot-experiments-report-daily-details.component.spec.ts
@@ -8,7 +8,7 @@ import { MockDotMessageService } from '@dotcms/utils-testing';
 
 import { DotExperimentsReportDailyDetailsComponent } from './dot-experiments-report-daily-details.component';
 
-import { DotExperimentsDetailsTableComponent } from '../../../shared/ui/dot-experiments-details-table/dot-experiments-details-table.component';
+import { DotExperimentsDetailsTableComponent } from '../../../../shared/ui/dot-experiments-details-table/dot-experiments-details-table.component';
 import { DotExperimentsReportsStore } from '../../store/dot-experiments-reports-store';
 
 const messageServiceMock = new MockDotMessageService({
